enhance test readability and change for loops iteration object to 8 bit unsigned 8 bits int type.

+++ software/acchess/src/board/board_comparator.cpp
@@ -1,4 +1,5 @@
 #include "board/board_comparator.hpp"
+#include <cstdint>
 
 namespace ac::chess {
 
@@ -10,8 +11,8 @@ std::vector<SquareChange> BoardComparator::compare(
     std::vector<SquareChange> changes;
     changes.reserve(4); // the number 4 here is chosen because the max number of squares that can change in a single chess move is 4 (castling).
 
-    for (int row = 0; row < 8; ++row) {
-        for (int col = 0; col < 8; ++col) {
+    for (uint8_t row = 0; row < 8; ++row) {
+        for (uint8_t col = 0; col < 8; ++col) {
             const Square square{row, col};
             const Piece before = previous.pieceAt(square);
             const Piece after = current.pieceAt(square);
